% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db_functions.R
\name{add_directory_path}
\alias{add_directory_path}
\title{Get Diagnosis Dates for List of ICD-9 diagnoses}
\usage{
add_directory_path(
  cond,
  group = "grant",
  base_path = NULL,
  years = 1:20,
  medicaid_years = 16:18,
  sources = c("ccae", "mdcr", "medicaid"),
  directory_path = NULL,
  overwrite = FALSE
)
}
\arguments{
\item{cond}{name of the condition}

\item{group}{name of group, options include 'grant', 'collab' (default is
'grant')}

\item{base_path}{base path for output, if not specified will use group defined
above}

\item{years}{years of ccae and mdcr data to collect}

\item{medicaid_years}{years of medicaid data to collect}

\item{sources}{options include ccae, mdcr, and medicaid, default is all three}

\item{directory_path}{path to where directory file to write to is located, if NULL
the defaul location will be used}

\item{overwrite}{logical indicator of whether to overwrite if 'cond' exists}
}
\description{
This function goes to the main database and pulls...
}
