% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ed_indicator.R
\name{add_ed_indicator}
\alias{add_ed_indicator}
\title{Add ED indicator}
\usage{
add_ed_indicator(data)
}
\arguments{
\item{data}{A tibble of outpatient keys}
}
\value{
A tibble of outpatient key including a source indicator for ED or outpatient
}
\description{
This function  identifies ED visits from outpatient visits based on specific criteria.
}
