% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ed_indicator.R
\name{add_ed_obs_indicator}
\alias{add_ed_obs_indicator}
\title{Add indicators for ED or observational stay}
\usage{
add_ed_obs_indicator(data)
}
\arguments{
\item{data}{A tibble of outpatient keys}
}
\value{
A tibble of outpatient key including a source indicator for ED, 
observational stay or outpatient
}
\description{
This function  identifies ED visits and observational stays from outpatient 
visits based on specific criteria.
}
