% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_patient_id_functions.R
\name{add_ed_obs_indicator_new}
\alias{add_ed_obs_indicator_new}
\title{Add indicators for ED or observational stay (New for patient_id)}
\usage{
add_ed_obs_indicator_new(data)
}
\arguments{
\item{data}{A tibble of outpatient keys}
}
\value{
A tibble of outpatient key including a source indicator for ED, 
observational stay or outpatient
}
\description{
This function  identifies ED visits and observational stays from outpatient 
visits based on specific criteria.
}
