% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_time_map.R
\name{assemble_time_map}
\alias{assemble_time_map}
\title{Assemble longitudinal timemap from small db}
\usage{
assemble_time_map(db_con, collect_tab = collect_table())
}
\arguments{
\item{db_con}{connection to the small database}

\item{collect_tab}{(optional) a collection table. This argument is only used to make temporary,
visit keys if no keys are found in the database}
}
\value{
a tibble containing the timemap
}
\description{
This function makes a longitudinal time_map and adds it to the small database. 
After running this function you can use `retrieve_time_map()` to collect the 
timemap from the small database.
}
