% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_map_objects.R
\name{build_dx_indicators}
\alias{build_dx_indicators}
\title{Build diagnosis indicators corresponding to a list of diagnoses}
\usage{
build_dx_indicators(
  condition_dx_list,
  db_con,
  collect_tab = collect_table(),
  return_keys_only = FALSE
)
}
\arguments{
\item{condition_dx_list}{a named list of icd9 and icd10 codes. Each condition should be named
and icd9_codes and icd10_codes should be stored as a named sub-list for each condition}

\item{db_con}{connection to the small database}

\item{collect_tab}{(optional) a collection table. This argument is only used to make temporary,
visit keys if no keys are found in the database}

\item{return_keys_only}{Logical to return only the  visit keys containing specific diagnosis codes}
}
\value{
a tibble containing keys and indicators for each condition
}
\description{
This function makes indicators corresponding to a list of diagnosis codes,
for visits identifiedby timemap keys. IF keys are not found in the database, optional
arguments allow keys to be generated temporarily (NOTE: this function cannot
be used to permanently add keys to the database, to permanently add keys use,
`add_time_map_keys()`).
}
