% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_final_time_map_function_parallel.R
\name{build_final_time_map_parallel}
\alias{build_final_time_map_parallel}
\title{Builds a time map containing visit specific information for a specific condition and cohort (in parallel).}
\usage{
build_final_time_map_parallel(
  condition_short_name,
  duration_prior_to_index = 365L,
  num_cores = NULL,
  cohort_path,
  ssd_list,
  collect_tab = collect_table()
)
}
\arguments{
\item{condition_short_name}{The condition of interest. Specifically the short_name version of the condition}

\item{duration_prior_to_index}{How far from the index should the time map extend back (e.g. 90, 180, 365)}

\item{num_cores}{The number of worker cores to use. If not specified will determined the number of cores based on the which ever
is the smallest value between number of rows in for collect_tab or detected number of cores - 1}

\item{cohort_path}{Path to the specific cohort for the condition of interest}

\item{ssd_list}{A list of diagnosis codes of interest. The diagnosis codes need to be seperated into
diagnosis categories (e.g. cough, fever, etc.). Additionally, within the categories, diagnosis codes should be seperated into ICD 9 and
ICD 10 specific codes, with list elements labled as icd9_codes and icd10_codes}

\item{collect_tab}{a collection table to build time_map from. By default all years will be used}
}
\value{
A tibble with all the specified diagnosis codes and the corresponding visit key where the diagnosis codes appeared
}
\description{
Builds a time map containing visit specific information for a specific condition and cohort (in parallel).
}
\examples{
ssds <- read_rds("/Shared/Statepi_Diagnosis/params/ssd_codes/ssd_hsv_enceph.RDS")

final_time_map <- build_final_time_map(condition_short_name = "hsv_enceph",
                                       duration_prior_to_index = 365L,
                                       cohort_path = "/Shared/Statepi_Diagnosis/grant_projects/hsv_enceph/scripts/validation/enrolled_ge_365/updated_index_cohort.RDS",
                                       ssd_list = ssds)
}
