% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_time_map_objects.R
\name{build_rx_indicators}
\alias{build_rx_indicators}
\title{Build rx indicators corresponding to a list of medications}
\usage{
build_rx_indicators(rx_list, db_con, collect_tab = collect_table())
}
\arguments{
\item{rx_list}{a named list of ndc codes. Each condition should be named and
the ndc codes should be stored as a named sub-list for each condition}

\item{db_con}{connection to the small database}

\item{collect_tab}{(optional) a collection table. This argument is only used to make temporary,
visit keys if no keys are found in the database}
}
\value{
a tibble containing keys and indicators for each condition
}
\description{
This function makes indicators corresponding to a list of ndc codes,
for visits identified by timemap keys. If keys are not found in the database,
this function will not run
`add_time_map_keys()`).
}
