% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_time_map.R
\name{build_time_map}
\alias{build_time_map}
\title{Build longitudinal timemap from small db}
\usage{
build_time_map(db_con, collect_tab = collect_table())
}
\arguments{
\item{db_con}{connection to the small database}

\item{collect_tab}{(optional) a collection table. This argument is only used to make temporary,
visit keys if no keys are found in the database}
}
\value{
a tibble containing the timemap
}
\description{
This function makes a time_map from the inpatient and outpatient visit keys,
contained in a small database. IF keys are not found in the database, optional
arguments allow keys to be generated temporarily (NOTE: this function cannot
be used to permanently add keys to the database, to permanently add keys use,
`add_time_map_keys()`).
}
