% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_time_map_parallel.R
\name{build_time_map_parallel}
\alias{build_time_map_parallel}
\title{Build longitudinal timemap from small db (in parallel)}
\usage{
build_time_map_parallel(
  collect_tab = collect_table(),
  db_path,
  db_con,
  num_cores = 10
)
}
\arguments{
\item{collect_tab}{(optional) a collection table. This argument is only used to make temporary,
visit keys if no keys are found in the database}

\item{db_path}{Path to the database}

\item{db_con}{connection to the small database}

\item{num_cores}{The number of worker cores to use. If not specified will use 10. If NULL is specified then it will determine the number of cores based on the which ever
is the smallest value between number of rows in for collect_tab or detected number of cores - 1}
}
\value{
a tibble containing the timemap
}
\description{
This function makes a time_map from the inpatient and outpatient visit keys,
contained in a small database. IF keys are not found in the database, optional
arguments allow keys to be generated temporarily (NOTE: this function cannot
be used to permanently add keys to the database, to permanently add keys use,
`add_time_map_keys_parallel()`).
}
