% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_patient_id_functions.R
\name{build_tm_keys}
\alias{build_tm_keys}
\title{Build tm (time_map) keys (new)}
\usage{
build_tm_keys(db_con)
}
\arguments{
\item{db_con}{a connection to a database}
}
\value{
A list containing three tibbles (timemap, outpatient keys, and inpatient keys)
}
\description{
This function goes to the database and assembles all of the inpatient an outpatient visits
(or just the visits specified by a collection table). It then assembles them into a logintudinal
time_map of distinct visits (defined by admission/service date and stdplac). Finally, each
visit is given a distinct visit key. The time_map and the distinct inpatient and outpatient visit keys
are returned as a list. NOTE: This function should generally only be run once after the small DB is
build, and it generally should not be called directly. Use the function `add_tm_keys()` to add the
time_map keys to the database.
}
