% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_patient_id_functions.R
\name{collect_plan}
\alias{collect_plan}
\title{Build collection plan}
\usage{
collect_plan(db_con)
}
\arguments{
\item{db_con}{a database connection}
}
\value{
a tibble
}
\description{
`collect_plan` scans a small db and sets up a collection tibble for holding 
and distributing queries across a rand of tables in the database
}
