% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_tables.R
\name{collect_table}
\alias{collect_table}
\title{Build collection tibble}
\usage{
collect_table(
  settings = c("inpatient", "outpatient", "rx"),
  sources = c("ccae", "mdcr"),
  years = 1:20
)
}
\arguments{
\item{settings}{"inpatient", "outpatient" or c("inpatient","outpatient")}

\item{sources}{"ccae" or "mdcr" or c("ccae","mdcr")}

\item{years}{vector of years to collect. Note: if no argument is provided
all years will be selected}
}
\value{
a tibble
}
\description{
`collect_table` sets up a collection tibble for holding and distributing
queries across a rand of tables in the database
}
\examples{
collect_table()
}
