% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egeoloc_labels}
\alias{egeoloc_labels}
\title{Geographic Location of Employee (EGEOLOC)}
\format{
A data frame with 66 rows and 3 variables:
\describe{
  \item{egeoloc}{Integer Value for geographic location}
  \item{label}{Geographic location Label}
  \item{level}{Indicator of label level, options include: `nation`, `region`, `state`, `city`}
}
}
\source{
Marketscan Documentation
}
\usage{
egeoloc_labels
}
\description{
A dataset containing labels for geographic location of employee
}
\keyword{datasets}
