% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_functions.R
\name{gather_enroll_data}
\alias{gather_enroll_data}
\title{Get enrollment data over multiple "enrollees" tables}
\usage{
gather_enroll_data(
  collect_tab = collect_table(),
  enrolid_list,
  collect_n = Inf,
  vars = c("dobyr", "sex"),
  db_con
)
}
\arguments{
\item{collect_tab}{A tibble with the specific setting (i.e. inpatient or outpatient), source (i.e. ccae or mdcr), and year to access
Default is all possible combinations of setting, source, and year}

\item{enrolid_list}{A list of enrolids for which enrollment data will be collected}

\item{collect_n}{The number of observations to return
is the smallest value between number of rows in for collect_tab or detected number of cores - 1}

\item{vars}{Vector of specific variables of interest in the "enrollees" tables (e.g. c("dobyr", "sex"))}

\item{db_con}{A connection to a database}
}
\value{
A tibble with information on each enrolid in enrolid_list. The number of columns in the tibble depends on
the number of vars selected
}
\description{
Get enrollment data over multiple "enrollees" tables
}
\examples{

out <- gather_enroll_data(enrolid_list = final_cohort$enrolid, db_con = db_con, vars = c("dobyr", "sex"))

}
