% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_get.R
\name{get_all_inpatient_dx_visits}
\alias{get_all_inpatient_dx_visits}
\title{Collect all inpatient diagnosis visits}
\usage{
get_all_inpatient_dx_visits(
  source,
  year,
  dx_num = TRUE,
  db_con,
  collect_n = Inf,
  keys
)
}
\arguments{
\item{source}{ccae, mdcr or medicaid}

\item{year}{year (as integer value)}

\item{dx_num}{a logical indicator of whether to collect diagnosis number. Default is TRUE.}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}

\item{keys}{what keys to return diagnoses for}
}
\value{
A tibble of variables from the respective table
}
\description{
Collect all inpatient visits along with the diagnosis order
}
