% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_get.R
\name{get_core_data}
\alias{get_core_data}
\title{Collect core data}
\usage{
get_core_data(setting, source, year, vars = c(), db_con, collect_n = Inf)
}
\arguments{
\item{setting}{inpatient or outpatient}

\item{source}{ccae or mdcr}

\item{year}{year (as integer value)}

\item{vars}{a vector of variables to collect. If NULL all variables in a table will be returned}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\value{
A tibble of variables from the respective core table
}
\description{
collect data from the core table
}
