% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_get_parallel.R
\name{get_core_data_parallel}
\alias{get_core_data_parallel}
\title{Gather variables from core information for a specific table (in parallel)}
\usage{
get_core_data_parallel(table_name, vars = c(), db_path, collect_n = Inf)
}
\arguments{
\item{table_name}{The specific core table name in the database to access}

\item{vars}{Vector of variables it the core table to return}

\item{db_path}{Path to the database}

\item{collect_n}{The number of observations to return}
}
\value{
A tibble with all the specified variables to return. Number of rows of tibble returned is determined by the collect_n argument
}
\description{
Gather variables from core information for a specific table (in parallel)
}
