% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_functions.R
\name{get_enroll_data}
\alias{get_enroll_data}
\title{Get enrollment data from one specific multiple "enrollees" table}
\usage{
get_enroll_data(
  table,
  enrolid_list,
  collect_n = Inf,
  vars = c("dobyr", "sex"),
  db_path
)
}
\arguments{
\item{table}{A tibble with a single row containing the specific source (i.e. ccae or mdcr) and year to access}

\item{enrolid_list}{A list of enrolids for which enrollment data will be collected}

\item{collect_n}{The number of observations to return}

\item{vars}{Vector of specific variables of interest in the "enrollees" tables (e.g. c("dobyr", "sex"))}

\item{db_path}{Path to the database}
}
\value{
A tibble with information on each enrolid in enrolid_list. The number of columns in the tibble depends on
the number of vars selected
}
\description{
Get enrollment data from one specific multiple "enrollees" table
}
