% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_index_get.R
\name{get_index_dx_visits}
\alias{get_index_dx_visits}
\title{Collect all diagnoses that occur in proximity to an index date}
\usage{
get_index_dx_visits(
  source,
  year,
  enrolid_table,
  days_before,
  days_after,
  db_con,
  collect_n = Inf
)
}
\arguments{
\item{source}{ccae, mdcr or medicaid}

\item{year}{year (as integer value)}

\item{enrolid_table}{a table containing enrolid and index_date}

\item{days_before}{an integer for number of days before index date to look}

\item{days_after}{an integer for number of days after index date to look}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\value{
A tibble of variables from the respective table
}
\description{
Collect all inpatient visits along with the diagnosis order
}
