% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db_functions.R
\name{get_linked_all_truven}
\alias{get_linked_all_truven}
\title{Get Linked All}
\usage{
get_linked_all_truven(
  table_name,
  source,
  year,
  join_table,
  join_by = "enrolid",
  con,
  collect_n = 10
)
}
\arguments{
\item{table_name}{name of table to pull}

\item{source}{"ccae" or "mdcr" or c("ccae","mdcr")}

\item{year}{vector of years to collect. Note: if no argument is provided
all years will be selected}

\item{join_table}{a table containing enrolids (or other) to join on}

\item{join_by}{a vector of variables to join by in the join table}

\item{con}{a database connection}

\item{collect_n}{number of rows to collect}
}
\value{
A list containing three tibbles (timemap, outpatient keys, and inpatient keys)
}
\description{
This function goes to the main database and pulls...
}
