% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_db_functions.R
\name{get_linked_visits_all}
\alias{get_linked_visits_all}
\title{Get Linked Visits (All)}
\usage{
get_linked_visits_all(
  setting,
  source,
  year,
  enrolid_table,
  con,
  collect_n = 10,
  HCUP = FALSE
)
}
\arguments{
\item{setting}{"inpatient", "outpatient", "rx"}

\item{source}{"ccae" or "mdcr" or c("ccae","mdcr")}

\item{year}{vector of years to collect. Note: if no argument is provided
all years will be selected}

\item{enrolid_table}{a table containing enrolids to pull data for}

\item{con}{a database connection}

\item{collect_n}{number of rows to collect}

\item{HCUP}{a logical indicator of whether collecting data from HCUP (vs Truven)}
}
\value{
A list containing three tibbles (timemap, outpatient keys, and inpatient keys)
}
\description{
This function goes to the main database and pulls...
}
