% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_get.R
\name{get_proc_dates}
\alias{get_proc_dates}
\title{Get Procedure visits for a set of procedure codes}
\usage{
get_proc_dates(
  source,
  year,
  proc_codes,
  setting = "inpatient",
  tbl_vars = NULL,
  db_con,
  collect_n = Inf
)
}
\arguments{
\item{source}{ccae, mdcr or medicaid}

\item{year}{year}

\item{proc_codes}{vector of procedure codes to pull visits for}

\item{setting}{"inpatient" or "outpatient"}

\item{tbl_vars}{names of variables to pull from corresponding tables. If NULL all tables}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\description{
This function collects dates for procedures
}
