% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_get.R
\name{get_rx_data}
\alias{get_rx_data}
\title{Get RX visits for a set of NDC codes}
\usage{
get_rx_data(source, year, ndc_codes, rx_vars = c(), db_con, collect_n = Inf)
}
\arguments{
\item{source}{ccae or mdcr}

\item{year}{year (as integer value)}

\item{ndc_codes}{a vector of ndc codes to lookup}

\item{rx_vars}{variables to collect from the rx table. If no variable names are provided
all variables in the rx table are returned. Default is rx_vars = c()}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\description{
This function collects records that correspond to a set of NDC codes
}
