% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_get.R
\name{get_rx_dates}
\alias{get_rx_dates}
\title{Get RX visits for a set of NDC codes}
\usage{
get_rx_dates(source, year, db_con, collect_n = Inf)
}
\arguments{
\item{source}{ccae or mdcr}

\item{year}{year (as integer value)}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\description{
This function collects dates for RX fills
}
