% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_get.R
\name{get_table_data}
\alias{get_table_data}
\title{Collect table data}
\usage{
get_table_data(table, source, year, vars = NULL, db_con, collect_n = Inf)
}
\arguments{
\item{table}{table to collect variables from}

\item{source}{ccae, mdcr or medicaid}

\item{year}{year (as integer value)}

\item{vars}{a vector of variables to collect. If NULL all variables in a table will be returned}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\value{
A tibble of variables from the respective table
}
\description{
Collect data from a generic table
}
