% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether.R
\name{gether_all_dx_keys}
\alias{gether_all_dx_keys}
\title{Gather all visit keys associated with a set of visits for a particular diagnosis}
\usage{
gether_all_dx_keys(collect_tab = collect_table(), db_con, keys)
}
\arguments{
\item{collect_tab}{A collection table}

\item{db_con}{A connection to the database}
}
\description{
This function return a tibble of
}
