% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether.R
\name{gether_all_inpatient_dx_keys}
\alias{gether_all_inpatient_dx_keys}
\title{Gather all visit keys associated with a set of visits}
\usage{
gether_all_inpatient_dx_keys(
  collect_tab = collect_table(),
  dx_num = TRUE,
  db_con,
  keys,
  primary = FALSE
)
}
\arguments{
\item{collect_tab}{A collection table}

\item{dx_num}{a logical indicator of whether to collect diagnosis number. Default is TRUE.}

\item{db_con}{A connection to the database}

\item{primary}{whether or not to only return primary diagnoses, if TRUE forces dx_num to TRUE}
}
\description{
This function return a tibble of
}
