% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether.R
\name{gether_core_data}
\alias{gether_core_data}
\title{Gather (gether) core data from all tables defined in a collection tab}
\usage{
gether_core_data(
  collect_tab = collect_table(),
  vars = c(),
  db_con,
  collect_n = Inf
)
}
\arguments{
\item{collect_tab}{a collection table}

\item{vars}{variables to collect}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\description{
Gather (gether) core data from all tables defined in a collection tab
}
