% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether_parallel.R
\name{gether_core_data_parallel}
\alias{gether_core_data_parallel}
\title{Gather variables from core information over multiple core tables (in parallel)}
\usage{
gether_core_data_parallel(
  collect_tab = collect_table(),
  vars = c(),
  db_path,
  collect_n = Inf,
  num_cores = NULL
)
}
\arguments{
\item{collect_tab}{A tibble with the specific setting (i.e. inpatient or outpatient), source (i.e. ccae or mdcr), and year to access.
Default is all possible combinations of setting, source, and year}

\item{vars}{Vector of variables it the core table to return}

\item{db_path}{Path to the database}

\item{collect_n}{The number of observations to return}

\item{num_cores}{The number of worker cores to use. If not specified will determined the number of cores based on the which ever
is the smallest value between number of rows in for collect_tab or detected number of cores - 1}
}
\value{
A tibble with all the specified variables to return over all designated core tables
}
\description{
Gather variables from core information over multiple core tables (in parallel)
}
