% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether.R
\name{gether_dx_keys}
\alias{gether_dx_keys}
\title{Gather visit keys associated with a set of visits for a particular diagnosis}
\usage{
gether_dx_keys(collect_tab = collect_table(), dx_list, db_con)
}
\arguments{
\item{collect_tab}{A collection table}

\item{dx_list}{A named list of icd9 and icd10 codes, with names of "icd9_codes" and
"icd10_codes", respectively}

\item{db_con}{A connection to the database}
}
\description{
This function return a tibble of
}
