% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether_large_db.R
\name{gether_dx_keys_delay_large_DB}
\alias{gether_dx_keys_delay_large_DB}
\title{Gather all visit keys containing specific diagnosis codes for large databases}
\arguments{
\item{collect_tab}{A tibble with the specific setting (i.e. inpatient or outpatient), source (i.e. ccae or mdcr), and year to access.
Default is all possible combinations of setting, source, and year}

\item{dx_list}{A list of specific diagnosis codes that are of interest. The diagnosis codes need to be seperated into into ICD 9 and
ICD 10 specific codes. The list elements need to be labled as icd9_codes and icd10_codes}

\item{db_path}{Path to the database}
}
\value{
A tibble with all the specified diagnosis codes and the corresponding visit key where the diagnosis codes appeared
}
\description{
Gather all visit keys containing specific diagnosis codes for large databases
}
