% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_index_gether.R
\name{gether_index_dx_visits}
\alias{gether_index_dx_visits}
\title{Gather (gether) dx data in proximity to index}
\usage{
gether_index_dx_visits(
  collect_tab = collect_table(),
  enrolid_table,
  days_before,
  days_after,
  db_con,
  collect_n = Inf
)
}
\arguments{
\item{collect_tab}{a collection table}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}

\item{vars}{variables to collect}
}
\description{
Gather (gether) dx data in proximity to index
}
