% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether.R
\name{gether_inpatient_dx_keys}
\alias{gether_inpatient_dx_keys}
\title{Gather visit keys associated with a set of visits for a particular inpatient diagnosis}
\usage{
gether_inpatient_dx_keys(
  collect_tab = collect_table(),
  dx_list,
  dx_num = TRUE,
  db_con
)
}
\arguments{
\item{collect_tab}{A collection table}

\item{dx_list}{A named list of icd9 and icd10 codes, with names of "icd9_codes" and
"icd10_codes", respectively}

\item{dx_num}{a logical indicator of whether to collect diagnosis number. Default is TRUE.}

\item{db_con}{A connection to the database}
}
\description{
This function return a tibble of
}
