% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether.R
\name{gether_proc_keys}
\alias{gether_proc_keys}
\title{Gather visit keys associated with a set of visits for a particular diagnosis}
\usage{
gether_proc_keys(collect_tab = collect_table(), proc_list, db_con)
}
\arguments{
\item{collect_tab}{A collection table}

\item{proc_list}{A named list of icd9, icd10 and CPT procedure codes, with names of "icd9_codes" and
"icd10_codes", respectively}

\item{db_con}{A connection to the database}
}
\description{
This function return a tibble of
}
