% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether.R
\name{gether_rx_dates}
\alias{gether_rx_dates}
\title{Gather (gether) rx dates from all rx tables defined in a collection tab}
\usage{
gether_rx_dates(collect_tab = collect_table(), db_con, collect_n = Inf)
}
\arguments{
\item{collect_tab}{a collection table}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\description{
Gather (gether) rx dates from all rx tables defined in a collection tab
}
