% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_tables_gether.R
\name{gether_rx_keys}
\alias{gether_rx_keys}
\title{Gather (gether) rx data from all tables defined in a collection tab}
\usage{
gether_rx_keys(
  collect_tab = collect_table(),
  ndc_codes,
  rx_vars = c("enrolid", "ndcnum", "svcdate"),
  db_con,
  collect_n = Inf
)
}
\arguments{
\item{collect_tab}{a collection table}

\item{ndc_codes}{a vector of ndc_codes to lookup}

\item{rx_vars}{variables to collect. Default is rx_vars = c("enrolid","ndcnum","svcdate")}

\item{db_con}{a connection to a database,}

\item{collect_n}{the number of rows to collect}
}
\description{
Gather (gether) rx data from all tables defined in a collection tab
}
