% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setting_type_label.R
\name{setting_type_labels}
\alias{setting_type_labels}
\title{Converts setting type integer values to their corresponding character forms.}
\usage{
setting_type_labels(setting_type_vector)
}
\arguments{
\item{setting_type_vector}{A vector of setting types in integer form}
}
\value{
A vector of setting types converted to the respective character forms
}
\description{
Converts setting type integer values to their corresponding character forms.
}
