% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stdplac_labels}
\alias{stdplac_labels}
\title{Place of Service Labels (STDPLAC)}
\format{
A data frame with 55 rows and 2 variables:
\describe{
  \item{stdplac}{Integer Value for stdplac}
  \item{label}{Place of Servce Label}
}
}
\source{
Marketscan Documentation
}
\usage{
stdplac_labels
}
\description{
A dataset containing labels for Place of Service (STDPLAC)
}
\keyword{datasets}
